Financial_Dtata_Analyser/pyproject.toml
[project]
name = "project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
